+++ CHANGELOG.md
@@ -29,6 +29,7 @@ For more information on SemVer, please visit http://semver.org/.
 
 #### Bugs
 
+* Fixed potential wrong URL on Ajax request in `WPDKListTableViewController`
 * Fixed `WPDKRoles`
 * Fixed potential bugs on filter `post_updated_messages` (for missing params) in Custom Post Type